requirements.txt
fastapi[all]
motor
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pyrogram
tgcrypto
pymongo[srv]
pydantic-settings
sse-starlette